// App/Models/FplDraftLeaguePlayerStatuses.ts
// Generated by https://quicktype.io

export interface FplDraftLeaguePlayerStatuses {
    element_status: ElementStatus[];
}

export interface ElementStatus {
    element:           number;
    in_accepted_trade: boolean;
    owner:             number | null;
    status:            Status;
}

export enum Status {
    A = "a",
    O = "o",
}